support user authentication in interfaces
A new Database object will provide the currently logged user as a
PersonModel, and through it, all related information. As a
consequence, most classes are now free of any global
state (e.g. reading/writing from a single global GroupModel).

Many bug fixes and layout changes have followed as a consequence of
adapting to the changes and testing them.

Preliminary Firebase support was added (for now, just a dependency in
build.gradle).