Registration of a new User
I've implemented the possibility to register with another button "register_button"", and inside the implementation is also called the function to send an email verification to the user. An ad-hoc dialog is showed in the case the user try to sign-in but he hasn't confirmed the mail yet.
Only when the user confirmed the mail , is possible for him to sign-in correctly , and therefore visualize the main activity of easyplit.