Bug Fixed
Thanks to @Flavio that observed a bug during the invitation of a new user.
The problem was exactly in the SUPER BIG SHIT api of google.
I mean , i discovered that when you call the function auth.createUserWithEmailAndPassword , a part of creating the user , after the function succeeds the new user is automatically signed in.
And of course it has no sense , it is a detail of this function.
By the way , i solved this problem by using the SharedPreferences Class into Android.
It basically saves login information of the user at the beginning, in a private mode inside the phone. In this way after i've created the new temporary user , i can sign in with the old credential. With-out encounter any more in this bug.