Fixed Bugs
There was the problem that when you clicked on SIgn in , everytime a new user was created (in the db) although it was already signed in for the first time (in the past). So for the moment , because of the Expiration date is today, i moved the actualizeOnDb on the Register-Phase , in this way the actualization of user information into the db is made only once. (Except for the stupid user that will click two times on Register , but we will solve it later on).