Fixed Bugs
In AddExpenses , there was the problem about the currency, because the USD actually is not working, now it will be present in the choice list , but as second choice (for further on bug fixing related). At the moment the first choice will be EUR that is the only one is really working for the moment.