[Bg Fixed] - AddExpenses MembersCheckList
I was getting crazy about 2 days , i tried several strategies about adding elements to listeview.
And all of the time the checking by debug was noticing me for correct behaviour.
So after my brain-fucking-time , i arrived to the conclusion that the problem was the graphical thread or stuff related.
In fact i was right.
The problem is a detail of the listView, in practice if you use a listView you cant put it inside a scrollView. Otherwise il will encounter a miss-behaviour , and on the screen will appear only one element although inside the listView actually there are more then one Element.