ExpenseDetailsActivity - Implemented
I 've put the listener in the right way. And updated the list-view, by using a proper adapter.
The data showed in this screen is eventListened only with the same instant in which the user press on the Expense.
If a change happens , it will be not listened instantaneously , but will be recognized by the user , only the next time he will press again on the same expense.
I opted for this solution because:
1) i considered is not necessary to make this screen real-time updated.
2) it is not a real-time log , but just an expense-detail.
3) the user for sure will not update a single expense every seconds.
4) Of course in this way the code becomes more readable and simple.