Bug Fixed: Crashing with dot-Numbers Fixed
During a function recompute inside GroupBalance , suddenly appeared an arithmetic-runtime-exception.
The reason was not considering the RoundingMode and also not considering the precision after dot , for decimal numbers.
It was simply solved by specifying this stuff inside Money Class.
Exactly in the point in which the MoneyClass is used For division.