Feature Navigation UpBack Button Implemented + Graphic Bug Fixed
The feature built-in in android by declaring in manifest file the parent activity , in this project is not available. Because we are working on a project where the base API is 15.
So i found out another solution.
I've set an home button in every single activity , in this way the user can go back to the main Group Activity every time he wants.
I have found some problems in the implementation of this button inside CreationGroup, because the graphic was a very big mess. So i fixed the graphic at first and then i was able to implement also here in the right way the button. (Of course i didn't touch the logic, it was not my business).
For the addExpenseActivity , i found some problems about to say to the compilator "dont go back to addExpenseActivity after you insert a new one Expense"", because of course after we have added an expense we don't want to come back to the AddExpense Screen but  instead to the GroupDetailsScreen. I fixed also this. Enjoy.