Refactor of AddExpense with the new Firebase-DB
I've inserted inside the relative group , a duplicate of expense.And i left also the old representation of expense inside table expenses into Firebase-DB. In this way it was not necessary change the Logic of ExpenseDetailsActivity, that was already pretty good. But this little change will be very useful to improve the performance of groupBalance Computation (That will require less call back to firebase).