Refactor of GroupBalance with the new Firebase-DB
Thanks to the previous commits and the previous refactoring of DB , i've improved the reading and computation of this task. In this way , everytime we need to recompute groupBalance we'll avoid to read the entire DB of firebase, but wil be enough to listen on changes of expenses inside the relative group.