Feature Implemented - Payer is a Financier
We have forgotten the case in which the payer is only a financier of the expense. It's an important feature , about the correctness of the groupBalance.
In practice for each single expense i check also if the payer in not involved in the payment of the expense , it means he is a financier. In this way we have to update correctly his residue.