Feature of GroupBalanceModel.java and ExpenseDetailsActivity.java improved - And Bug Fixed on AddExpenses.java
Now the group Balance and the Expense Detail has the precision of single cent, in the calculation. It means that now is also considered the case in which the division of the amount in not exact with respect to the number of members.

In Add Expenses is now fixed the bug about the approximation, before this intervention there were too much numbers after the dot.

In GroupBalanceAdapter i added the new String , that is visualized in the layout list_item_member.xml
In order we'll be able (later on) to associate to each member involved in the payment a hint about who has to give back the money to who.

In GroupBalance i also create a model for the function that will be used to support and to update properly the new String created on the layout list_item_member.
For the moment is only showed that the member with negative balance has to give back and instead the member with a positive balance has to receive.