Graphic Hints About Creditors and Debtors
The graphic has been updated, basically taking in consideration the previous commit where there is all the logic behind.

Now in GroupBalance is possible to see who has to give back money to. In this way , the User interaction will be simplified a lot. Specially when later on will be implemented also the feature of the payment.

The concept of a catchUpGroup is basically the concept where for every creditor will be a list of debtors that for sure by adding their debts , and summing the total amount with the residue of the creditor , the final result will be zero.