Visualization of expenses by Date descending order
Of course after the merge with master , all new expenses added by the new version of Addexpense will work, instead the old-ones added before in other groups will crash. But This problem will be solved automatically when we will destroy all old ones expenses, and will replace them with the new ones.

P.S. This Feature will of course entails into a refactor of the ModifyExpenses implemented by Sebastiano, but it will be not difficult to modify because its just a field to add "timestamp_number"