Fixed Little Bug on Group Balance
The problem was , that during the constructions of debtors it was consider as a debtor also who was to owes 0.00 , instead we have to consider as a debtor only if its residue is less then 0.
In order to appear as a debtor in the groupBalance Screen , only if he is really a debtor, otherwise not.