Temporary Feature On Creation Group Deactivated
The possibility of adding multiple users directly from Creation Group had cause some glitch problems in Group Screen, probably because a bad managing of thread in android.
For this reason , in order to give to Anna a stable version , where is always possible to add users to the group by invite inside the group it self, we decided to de-activate this functionality from CreationGroup Temporary.
Temporary because maybe @Flavio will find a fix for this problem.