Little Bug Fixed About Inserting Amount
There was a problem when you were using a version of language in your phone , respect another.
In particularly there was the problem that in italian the Decimal Separator is the comma instead in english is the point.
This cause number exception during the insertion of the amount ,so  if you change your system language the problem was due also to the wrong appearing of the wrong keyboard related to the location of your system language.

Now is fixed and it is working with English and with italian version.