Fixed Bug Related To The Crazy One Cent
Basically in TripWithFriends (And Also other group) , happens that if you put 5 euro , and change the payer, Sometimes it appears that a member with out any payment involved is included in the balance. And this bug was due to the fact , that in distribute rest we consider all members. Instead now we consider only members with a residue != 0 , and now the bug is fixed.