RollBack to Old DistributeRest Method in ExpenseDetails
TODO , when finally we'll fix the problem on groupBalance , we should try to make the rest distribution identical also in groupBalance. In this way , the user will see the same distribution also for One Expense in the group. Maybe the solution could be into making the order of members reading the same.