Fixed the Super Bug Of GroupBalance
The Logic of calculation was wrong in some points , and further on , there were also bad assignments, and there were also a problem of approximation.

Now everything seem working fine.

During my tests i discovered that if we use One single Currency for the group and the same Currency for the expenses of the group is very difficult that some catch_up groups are not let's say cent-correct (Because now is used a very precise approximation scale).

But instead if we mix the concept (i.e. Group currency of one type , and Expenses of different types respect to the Currency of The Group) it could happen that in the GroupBalanceVisualization (Because of the double way conversion , we could encounter an error of one cent in the catch_up groups)