Fixed Bug Of Concurrence Modification Of A Single Expense
The bug was that when i was pressing merge balance, while somebody else in a groupd was adding or modifying expenses or payents.It could happen that the progressBar get locked , by blocking the application.
Instead now by using a different approach , and all concurrent structures where is necessary, this but is no more detected.
Please if you have time Sebastiano , try also by your selfe if the bug is eliminated definetly.